requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy<2
wandb
pytorch-ignite
Pillow
